requirements.txt
asgiref==3.8.1
astroid==3.2.2
black==24.4.2
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
dill==0.3.8
dj-database-url==2.2.0
dj3-cloudinary-storage==0.0.6
Django==5.0.6
django-allauth==0.63.3
django-debug-toolbar==4.4.2
django-summernote==0.8.20.0
django-taggit==5.0.1
filetype==1.2.0
gunicorn==22.0.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2==2.9.9
pylint==3.2.5
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tomlkit==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.7.0
